corrección de error en resultado del costo de producción

+++ src/Library/Step.cs
@@ -8,7 +8,7 @@ namespace Full_GRASP_And_SOLID.Library
 {
     public class Step
     {
-        public Step(Product input, double quantity, Equipment equipment, int time)
+        public Step(Product input, double quantity, Equipment equipment, double time)
         {
             this.Quantity = quantity;
             this.Input = input;
@@ -20,7 +20,7 @@ namespace Full_GRASP_And_SOLID.Library
 
         public double Quantity { get; set; }
 
-        public int Time { get; set; }
+        public double Time { get; set; }
 
         public Equipment Equipment { get; set; }
     }
